requirements.txt
pytest >= 7.2.0
selene >= 2.0.b16
python-dotenv >= 0.21.0
allure-pytest >= 2.12.0